DAG DPOS Proposal (#620)
* Code format

* WIP in DAG DPOS proposal. Using genesis DPOS state for DAG proposal

* 1. Implement DPOS DAG blocks proposal and verification
2. Refactor DAG block, decoupling DAG block manager from DAG block structure

* Fix DPOS DAG block verification

* Move dag block manager into DAG module

* Fix merge conflicts

* Use deposit delay for DPOS period

* Small fix
