+++ src/CMakeLists.txt
@@ -96,6 +96,7 @@ set(SOURCES
         logger/logger_config.hpp
         logger/log.hpp
         chain/state_api.hpp
+        dag/dag_block_manager.hpp
         consensus/pbft_manager.hpp
 
         # ---- private sources -----
@@ -136,6 +137,7 @@ set(SOURCES
         logger/logger_config.cpp
         logger/log.cpp
         common/types.cpp
+        dag/dag_block_manager.cpp
         )
 
 add_subdirectory(network/rpc)
