+++ src/consensus/pbft_manager.hpp
@@ -32,7 +32,7 @@ class PbftManager {
 
   PbftManager(PbftConfig const &conf, std::string const &genesis, addr_t node_addr, std::shared_ptr<DbStorage> db,
               std::shared_ptr<PbftChain> pbft_chain, std::shared_ptr<VoteManager> vote_mgr,
-              std::shared_ptr<DagManager> dag_mgr, std::shared_ptr<BlockManager> blk_mgr,
+              std::shared_ptr<DagManager> dag_mgr, std::shared_ptr<DagBlockManager> dag_blk_mgr,
               std::shared_ptr<FinalChain> final_chain, std::shared_ptr<Executor> executor, secret_t node_sk,
               vrf_sk_t vrf_sk);
   ~PbftManager();
@@ -148,7 +148,7 @@ class PbftManager {
   std::shared_ptr<DagManager> dag_mgr_ = nullptr;
   std::shared_ptr<Network> network_ = nullptr;
   std::shared_ptr<TaraxaCapability> capability_ = nullptr;
-  std::shared_ptr<BlockManager> blk_mgr_;
+  std::shared_ptr<DagBlockManager> dag_blk_mgr_;
   std::shared_ptr<FinalChain> final_chain_;
   std::shared_ptr<Executor> executor_;
