+++ src/dag/dag_block.cpp
@@ -187,234 +187,4 @@ bytes DagBlock::rlp(bool include_sig) const {
 
 blk_hash_t DagBlock::sha3(bool include_sig) const { return dev::sha3(rlp(include_sig)); }
 
-BlockManager::BlockManager(vdf_sortition::VdfConfig const &vdf_config, size_t capacity, unsigned num_verifiers,
-                           addr_t node_addr, std::shared_ptr<DbStorage> db, std::shared_ptr<TransactionManager> trx_mgr,
-                           logger::Logger log_time, uint32_t queue_limit)
-    : capacity_(capacity),
-      num_verifiers_(num_verifiers),
-      blk_status_(10000, 100),
-      seen_blocks_(10000, 100),
-      queue_limit_(queue_limit),
-      db_(db),
-      trx_mgr_(trx_mgr),
-      log_time_(log_time),
-      vdf_config_(vdf_config) {
-  LOG_OBJECTS_CREATE("BLKQU");
-}
-
-BlockManager::~BlockManager() { stop(); }
-
-void BlockManager::start() {
-  if (bool b = true; !stopped_.compare_exchange_strong(b, !b)) {
-    return;
-  }
-  LOG(log_nf_) << "Create verifier threads = " << num_verifiers_ << std::endl;
-  verifiers_.clear();
-  for (auto i = 0; i < num_verifiers_; ++i) {
-    verifiers_.emplace_back([this, i]() { this->verifyBlock(); });
-  }
-}
-
-void BlockManager::stop() {
-  if (bool b = false; !stopped_.compare_exchange_strong(b, !b)) {
-    return;
-  }
-  cond_for_unverified_qu_.notify_all();
-  cond_for_verified_qu_.notify_all();
-  for (auto &t : verifiers_) {
-    t.join();
-  }
-}
-
-bool BlockManager::isBlockKnown(blk_hash_t const &hash) {
-  auto known = seen_blocks_.count(hash);
-  if (!known) return getDagBlock(hash) != nullptr;
-  return true;
-}
-
-std::shared_ptr<DagBlock> BlockManager::getDagBlock(blk_hash_t const &hash) const {
-  std::shared_ptr<DagBlock> ret;
-  auto blk = seen_blocks_.get(hash);
-  if (blk.second) {
-    ret = std::make_shared<DagBlock>(blk.first);
-  }
-  if (!ret) {
-    return db_->getDagBlock(hash);
-  }
-  return ret;
-}
-
-bool BlockManager::pivotAndTipsValid(DagBlock const &blk) {
-  auto status = blk_status_.get(blk.getPivot());
-  if (status.second && status.first == BlockStatus::invalid) {
-    blk_status_.update(blk.getHash(), BlockStatus::invalid);
-    LOG(log_dg_) << "DAG Block " << blk.getHash() << " pivot " << blk.getPivot() << " unavailable";
-    return false;
-  }
-  for (auto const &t : blk.getTips()) {
-    auto status = blk_status_.get(t);
-    if (status.second && status.first == BlockStatus::invalid) {
-      blk_status_.update(blk.getHash(), BlockStatus::invalid);
-      LOG(log_dg_) << "DAG Block " << blk.getHash() << " tip " << t << " unavailable";
-      return false;
-    }
-  }
-  return true;
-}
-
-level_t BlockManager::getMaxDagLevelInQueue() const {
-  level_t max_level = 0;
-  {
-    uLock lock(shared_mutex_for_unverified_qu_);
-    if (unverified_qu_.size() != 0) max_level = unverified_qu_.rbegin()->first;
-  }
-  {
-    uLock lock(shared_mutex_for_verified_qu_);
-    if (verified_qu_.size() != 0) max_level = std::max(verified_qu_.rbegin()->first, max_level);
-  }
-  return max_level;
-}
-
-void BlockManager::insertBlock(DagBlock const &blk) {
-  if (isBlockKnown(blk.getHash())) {
-    LOG(log_nf_) << "Block known " << blk.getHash();
-    return;
-  }
-  pushUnverifiedBlock(std::move(blk), true /*critical*/);
-  LOG(log_time_) << "Store cblock " << blk.getHash() << " at: " << getCurrentTimeMilliSeconds()
-                 << " ,trxs: " << blk.getTrxs().size() << " , tips: " << blk.getTips().size();
-}
-
-void BlockManager::pushUnverifiedBlock(DagBlock const &blk, std::vector<Transaction> const &transactions,
-                                       bool critical) {
-  if (queue_limit_ > 0) {
-    auto queue_size = getDagBlockQueueSize();
-    if (queue_limit_ > queue_size.first + queue_size.second) {
-      LOG(log_wr_) << "Warning: block queue large. Unverified queue: " << queue_size.first
-                   << "; Verified queue: " << queue_size.second << "; Limit: " << queue_limit_;
-    }
-  }
-  seen_blocks_.update(blk.getHash(), blk);
-  {
-    uLock lock(shared_mutex_for_unverified_qu_);
-    if (critical) {
-      blk_status_.insert(blk.getHash(), BlockStatus::proposed);
-
-      unverified_qu_[blk.getLevel()].emplace_front(std::make_pair(blk, transactions));
-      LOG(log_dg_) << "Insert unverified block from front: " << blk.getHash() << std::endl;
-    } else {
-      blk_status_.insert(blk.getHash(), BlockStatus::broadcasted);
-      unverified_qu_[blk.getLevel()].emplace_back(std::make_pair(blk, transactions));
-      LOG(log_dg_) << "Insert unverified block from back: " << blk.getHash() << std::endl;
-    }
-  }
-  cond_for_unverified_qu_.notify_one();
-}
-
-void BlockManager::insertBroadcastedBlockWithTransactions(DagBlock const &blk,
-                                                          std::vector<Transaction> const &transactions) {
-  if (isBlockKnown(blk.getHash())) {
-    LOG(log_dg_) << "Block known " << blk.getHash();
-    return;
-  }
-  pushUnverifiedBlock(blk, transactions, false /*critical*/);
-  LOG(log_time_) << "Store ncblock " << blk.getHash() << " at: " << getCurrentTimeMilliSeconds()
-                 << " ,trxs: " << blk.getTrxs().size() << " , tips: " << blk.getTips().size();
-}
-
-void BlockManager::pushUnverifiedBlock(DagBlock const &blk, bool critical) {
-  pushUnverifiedBlock(blk, std::vector<Transaction>(), critical);
-}
-
-std::pair<size_t, size_t> BlockManager::getDagBlockQueueSize() const {
-  std::pair<size_t, size_t> res;
-  {
-    uLock lock(shared_mutex_for_unverified_qu_);
-    res.first = 0;
-    for (auto &it : unverified_qu_) {
-      res.first += it.second.size();
-    }
-  }
-
-  {
-    uLock lock(shared_mutex_for_verified_qu_);
-    res.second = 0;
-    for (auto &it : verified_qu_) {
-      res.second += it.second.size();
-    }
-  }
-  return res;
-}
-
-DagBlock BlockManager::popVerifiedBlock() {
-  uLock lock(shared_mutex_for_verified_qu_);
-  while (verified_qu_.empty() && !stopped_) {
-    cond_for_verified_qu_.wait(lock);
-  }
-  if (stopped_) return DagBlock();
-
-  auto blk = verified_qu_.begin()->second.front();
-  verified_qu_.begin()->second.pop_front();
-  if (verified_qu_.begin()->second.empty()) verified_qu_.erase(verified_qu_.begin());
-  return blk;
-}
-
-void BlockManager::pushVerifiedBlock(DagBlock const &blk) {
-  uLock lock(shared_mutex_for_verified_qu_);
-  verified_qu_[blk.getLevel()].emplace_back(blk);
-}
-
-void BlockManager::verifyBlock() {
-  while (!stopped_) {
-    std::pair<DagBlock, std::vector<Transaction>> blk;
-    {
-      uLock lock(shared_mutex_for_unverified_qu_);
-      while (unverified_qu_.empty() && !stopped_) {
-        cond_for_unverified_qu_.wait(lock);
-      }
-      if (stopped_) return;
-      blk = unverified_qu_.begin()->second.front();
-      unverified_qu_.begin()->second.pop_front();
-      if (unverified_qu_.begin()->second.empty()) unverified_qu_.erase(unverified_qu_.begin());
-    }
-    auto status = blk_status_.get(blk.first.getHash());
-    // Verifying transaction ...
-    LOG(log_time_) << "VerifyingTrx block  " << blk.first.getHash() << " at: " << getCurrentTimeMilliSeconds();
-    // only need to verify if this is a broadcasted block (proposed block are
-    // generated by verified trx)
-    if (!status.second || status.first == BlockStatus::broadcasted) {
-      bool valid = trx_mgr_->verifyBlockTransactions(blk.first, blk.second);
-      if (valid) {
-        // Verify VDF solution
-        vdf_sortition::VdfSortition vdf = blk.first.getVdf();
-        if (!vdf.verifyVdf(vdf_config_, getRlpBytes(blk.first.getLevel()), blk.first.getPivot().asBytes())) {
-          LOG(log_er_) << "DAG block " << blk.first.getHash() << " failed on VDF verification with pivot hash "
-                       << blk.first.getPivot();
-          valid = false;
-        }
-      }
-      if (!valid) {
-        LOG(log_er_) << "Ignore block " << blk.first.getHash() << " since it has invalid or missing transactions";
-        blk_status_.update(blk.first.getHash(), BlockStatus::invalid);
-        continue;
-      }
-    }
-
-    {
-      uLock lock(shared_mutex_for_verified_qu_);
-      if (status.second && status.first == BlockStatus::proposed) {
-        verified_qu_[blk.first.getLevel()].emplace_front(blk.first);
-      } else if (!status.second || status.first == BlockStatus::broadcasted) {
-        verified_qu_[blk.first.getLevel()].emplace_back(blk.first);
-      }
-    }
-    blk_status_.update(blk.first.getHash(), BlockStatus::verified);
-
-    LOG(log_time_) << "VerifiedTrx stored " << blk.first.getHash() << " at: " << getCurrentTimeMilliSeconds();
-
-    cond_for_verified_qu_.notify_one();
-    LOG(log_dg_) << "Verified block: " << blk.first.getHash() << std::endl;
-  }
-}
-
 }  // namespace taraxa
