+++ src/network/network.hpp
@@ -17,7 +17,7 @@
 
 #include "config/config.hpp"
 #include "consensus/vote.hpp"
-#include "dag/dag_block.hpp"
+#include "dag/dag_block_manager.hpp"
 #include "taraxa_capability.hpp"
 #include "transaction_manager/transaction.hpp"
 #include "util/util.hpp"
@@ -33,12 +33,12 @@ class Network {
   Network(NetworkConfig const &config, std::string const &networkFile, std::string const &genesis, addr_t node_addr,
           std::shared_ptr<DbStorage> db, std::shared_ptr<PbftManager> pbft_mgr, std::shared_ptr<PbftChain> pbft_chain,
           std::shared_ptr<VoteManager> vote_mgr, std::shared_ptr<DagManager> dag_mgr,
-          std::shared_ptr<BlockManager> blk_mgr, std::shared_ptr<TransactionManager> trx_mgr, public_t node_pk,
+          std::shared_ptr<DagBlockManager> dag_blk_mgr, std::shared_ptr<TransactionManager> trx_mgr, public_t node_pk,
           uint32_t lambda_ms_min);
   Network(NetworkConfig const &config, std::string const &networkFile, secret_t const &sk, std::string const &genesis,
           addr_t node_addr, std::shared_ptr<DbStorage> db, std::shared_ptr<PbftManager> pbft_mgr,
           std::shared_ptr<PbftChain> pbft_chain, std::shared_ptr<VoteManager> vote_mgr,
-          std::shared_ptr<DagManager> dag_mgr, std::shared_ptr<BlockManager> blk_mgr,
+          std::shared_ptr<DagManager> dag_mgr, std::shared_ptr<DagBlockManager> dag_blk_mgr,
           std::shared_ptr<TransactionManager> trx_mgr, public_t node_pk, uint32_t lambda_ms_min);
   ~Network();
   void start(bool boot_node = false);
@@ -84,7 +84,7 @@ class Network {
   std::shared_ptr<PbftChain> pbft_chain_;
   std::shared_ptr<VoteManager> vote_mgr_;
   std::shared_ptr<DagManager> dag_mgr_;
-  std::shared_ptr<BlockManager> blk_mgr_;
+  std::shared_ptr<DagBlockManager> dag_blk_mgr_;
   std::shared_ptr<TransactionManager> trx_mgr_;
   public_t node_pk_;
   LOG_OBJECTS_DEFINE;
